Refactored integration tests to inject bootstrap sql file
