+++ lib/router/data.js
@@ -17,22 +17,21 @@
  */
 "use strict";
 
-var Insert = require('./lib/insert');
-var Select = require('./lib/select');
-var Update = require('./lib/update');
-var Delete = require('./lib/delete');
+var Insert = require('./../insert');
+var Select = require('./../select');
+var Update = require('./../update');
+var Delete = require('./../delete');
 var _ = require('lodash');
 
 module.exports = function(db, logger) {
-  var router = require('koa-router')();
+  var router = require('koa-router')({'prefix': '/data'});
 
-//var fetchData = require('./lib/fetch')(this.koa.context.db, this.logger);
   var insert = new Insert(db, logger);
   var select = new Select(db, logger);
   var update = new Update(db, logger);
   var delet = new Delete(db, logger);
 
-// TABLE
+  // TABLE
   router.param('table', function * checkTableIsNotInternal(table, next) {
     if (_.startsWith(table, 'pg_')) {
       throw Boom.forbidden('You cannot access internal tables.');
@@ -47,8 +46,8 @@ module.exports = function(db, logger) {
     yield next;
   });
 
-// POST
-  router.post('/data/:table', function* addDataHandler() {
+  // POST
+  router.post('/:table', function* addDataHandler() {
     if (!this.is('json')) {
       throw Boom.badRequest('Content-Type needs to be "application/json"');
     }
@@ -56,13 +55,13 @@ module.exports = function(db, logger) {
     this.body = yield insert.query(this.table, this.request.body);
   });
 
-// GET
-  router.get('/data/:table', function* fetchDataHandler() {
+  // GET
+  router.get('/:table', function* fetchDataHandler() {
     this.body = yield select.query(this.table, this.request.query);
   });
 
-// PATCH
-  router.patch('/data/:table', function* updateDatahandler() {
+  // PATCH
+  router.patch('/:table', function* updateDatahandler() {
     if (!this.is('json')) {
       throw Boom.badRequest('Content-Type needs to be "application/json"');
     }
@@ -71,7 +70,7 @@ module.exports = function(db, logger) {
   });
 
   // DELETE
-  router.delete('/data/:table', function* deleteDataHandler() {
+  router.delete('/:table', function* deleteDataHandler() {
     this.body = yield delet.query(this.table, this.request.query);
   });
