+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node standalone.js",
     "test": "mocha test/unit/lib/ --require test/unit/bootstrap.js",
-    "integration": "mocha test/integration/*.js --require test/integration/bootstrap.js -t 10000",
+    "integration": "mocha test/integration/data/*.js  test/integration/authentication/*.js --require test/integration/bootstrap.js -t 10000",
     "coverage": "./node_modules/istanbul-harmony/lib/cli.js cover node_modules/mocha/bin/_mocha -- test/unit/lib/* --require test/unit/bootstrap.js"
   },
   "config": {
