+++ test/integration/data/patch.js
@@ -17,7 +17,7 @@
  */
 "use strict";
 
-var test = new ManatiIntegrationTest();
+var test = new ManatiIntegrationTest(__dirname + '/../bootstrap.sql');
 const chance = require('chance').Chance();
 const async = require('async');
 const sprintf = require("sprintf-js").sprintf;
@@ -34,31 +34,37 @@ var checkLength = function(route, length) {
   };
 };
 
-describe('DELETE /data/:table', function (done) {
+describe('PATCH /data/:table', function (done) {
   before(function (done) {
     test.start(done);
   });
 
-  it('DELETE /data/uuid_data?uuid=199F5EFB-2DF6-42CF-90D7-61D90212C74A', function (done) {
+  it('PATCH /data/uuid_data?uuid=199F5EFB-2DF6-42CF-90D7-61D90212C74A', function (done) {
     var uuid = chance.guid();
 
     async.series([
-      function(cb) {test.app.delete('/data/uuid_data?uuid=199F5EFB-2DF6-42CF-90D7-61D90212C74A')
+      function(cb) {test.app.patch('/data/uuid_data?uuid=199F5EFB-2DF6-42CF-90D7-61D90212C74A')
         .set('Content-Type', 'application/json')
-        .expect(200, [{uuid: '199f5efb-2df6-42cf-90d7-61d90212c74a'}]).end(cb);
+        .send({
+          uuid: uuid
+        })
+        .expect(200, [{uuid: uuid}]).end(cb);
       },
-      checkLength('/data/uuid_data', 2)
+      checkLength('/data/uuid_data', 3)
     ], done);
   });
 
-  it('DELETE /data/number_data?smallint_number=gte::32767', function (done) {
+  it('PATCH /data/number_data?smallint_number=gte::32767', function (done) {
     async.series([
       function (cb) {
-        test.app.delete('/data/number_data?smallint_number=gte::32767')
+        test.app.patch('/data/number_data?smallint_number=gte::32767')
           .set('Content-Type', 'application/json')
+          .send({
+            'int_number': 1
+          })
           .expect(200, [{
             smallint_number: 32767,
-            int_number: 2147483647,
+            int_number: 1,
             bigint_number: '9223372036854775807',
             decimal_number: '1.2',
             numeric_number: '1.3',
@@ -67,7 +73,7 @@ describe('DELETE /data/:table', function (done) {
           }])
           .end(cb);
       },
-      checkLength('/data/number_data', 1)
+      checkLength('/data/number_data', 2)
     ], done);
   });
